+++ lang/fi/responses.php
@@ -106,6 +106,7 @@ return [
     'signer_not_found_or_already_signed' => 'Allekirjoittajaa ei löytynyt tai jo allekirjoitettu',
     'cannot_delete_signed_contract' => 'Allekirjoitettua sopimusta ei voi poistaa',
     'cannot_update_signed_contract' => 'Allekirjoitettua sopimusta ei voi päivittää',
-    'enable_js_to_use_app' => 'Olemme pahoillamme, mutta tämä sovellus ei toimi kunnolla ilman JavaScriptiä käytössä. Ota se käyttöön jatkaaksesi.',
+    'enable_js_to_use_app' =>
+        'Olemme pahoillamme, mutta tämä sovellus ei toimi kunnolla ilman JavaScriptiä käytössä. Ota se käyttöön jatkaaksesi.',
     'cannot_delete_own_account' => 'Omaa tiliä ei voi poistaa',
 ];